공장 CRUD
 상점 UD 권한 강화

+++ account-service/src/main/java/com/msa/account/domain/store/service/StoreService.java
@@ -5,36 +5,36 @@ import com.msa.account.domain.store.entity.AdditionalOption;
 import com.msa.account.domain.store.entity.Store;
 import com.msa.account.domain.store.repository.AdditionalOptionRepository;
 import com.msa.account.domain.store.repository.StoreRepository;
-import com.msa.account.global.domain.dto.AccountDto;
-import com.msa.account.global.domain.dto.CommonOptionDto;
+import com.msa.account.global.domain.dto.util.AuthorityUserRoleUtil;
 import com.msa.account.global.domain.entity.*;
 import com.msa.account.global.domain.repository.AddressRepository;
 import com.msa.account.global.domain.repository.CommonOptionRepository;
 import com.msa.account.global.domain.repository.GoldHarryRepository;
+import com.msa.account.global.exception.ExceptionMessage;
+import com.msa.account.global.exception.NotAuthorityException;
 import com.msa.account.global.exception.NotFoundException;
-import com.msacommon.global.jwt.JwtUtil;
 import com.msacommon.global.util.CustomPage;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Pageable;
 
-import java.util.Optional;
+import static com.msa.account.global.exception.ExceptionMessage.*;
 
 @Slf4j
 @Service
-@Transactional(readOnly = true)
+@Transactional
 public class StoreService {
 
-    private final JwtUtil jwtUtil;
+    private final AuthorityUserRoleUtil authorityUserRoleUtil;
     private final StoreRepository storeRepository;
     private final AddressRepository addressRepository;
     private final GoldHarryRepository goldHarryRepository;
     private final CommonOptionRepository commonOptionRepository;
     private final AdditionalOptionRepository additionalOptionRepository;
 
-    public StoreService(JwtUtil jwtUtil, StoreRepository storeRepository, AddressRepository addressRepository, GoldHarryRepository goldHarryRepository, CommonOptionRepository commonOptionRepository, AdditionalOptionRepository additionalOptionRepository) {
-        this.jwtUtil = jwtUtil;
+    public StoreService(AuthorityUserRoleUtil authorityUserRoleUtil, StoreRepository storeRepository, AddressRepository addressRepository, GoldHarryRepository goldHarryRepository, CommonOptionRepository commonOptionRepository, AdditionalOptionRepository additionalOptionRepository) {
+        this.authorityUserRoleUtil = authorityUserRoleUtil;
         this.storeRepository = storeRepository;
         this.addressRepository = addressRepository;
         this.goldHarryRepository = goldHarryRepository;
@@ -43,44 +43,28 @@ public class StoreService {
     }
 
     //상점 호출(info)
-    public AccountDto.accountInfo getStoreInfo(String storeId) {
-        AccountDto.accountInfo storeInfo = storeRepository.findByStoreId(Long.valueOf(storeId))
-                .orElseThrow(() -> new NotFoundException("대상을 찾을 수 없습니다."));
-
-        storeInfo.getTradeTypeTitle();
-        storeInfo.getLevelTypeLevel();
-
-        return AccountDto.accountInfo.builder()
-                .createAt(storeInfo.getCreateAt())
-                .businessName(storeInfo.getBusinessName())
-                .businessOwnerName(storeInfo.getBusinessOwnerName())
-                .businessNumber1(storeInfo.getBusinessNumber1())
-                .businessNumber2(storeInfo.getBusinessNumber2())
-                .faxNumber(storeInfo.getFaxNumber())
-                .address(storeInfo.getAddress())
-                .note(storeInfo.getNote())
-                .level(storeInfo.getLevel())
-                .tradeType(storeInfo.getTradeType())
-                .goldLoss(storeInfo.getGoldLoss())
-                .build();
+    @Transactional(readOnly = true)
+    public StoreDto.StoreSingleResponse getStoreInfo(String storeId) {
+        return storeRepository.findByStoreId(Long.valueOf(storeId))
+                .orElseThrow(() -> new NotFoundException(ExceptionMessage.NOT_FOUND_STORE));
     }
     //상점 목록 호출(infoList)
-    public CustomPage<AccountDto.accountInfo> getStoreList(Pageable pageable) {
+    @Transactional(readOnly = true)
+    public CustomPage<StoreDto.StoreResponse> getStoreList(Pageable pageable) {
         return storeRepository.findAllStore(pageable);
     }
 
     //상점 추가
-    @Transactional
     public void addStore(StoreDto.StoreRequest storeInfo) {
         
         //상점 이름 검증
         if (storeRepository.existsByStoreName(storeInfo.getStoreInfo().getStoreName())) {
-            throw new NotFoundException("이미 존재하는 상점입니다.");
+            throw new NotFoundException(NOT_FOUND_STORE);
         }
 
         //로스
         GoldHarry goldHarry = goldHarryRepository.findById(Long.valueOf(storeInfo.getCommonOptionInfo().getGoldHarryId()))
-                .orElseThrow(() -> new NotFoundException("잘못된 해리값 입력입니다."));
+                .orElseThrow(() -> new NotFoundException(WRONG_HARRY));
 
         Store newstore = storeInfo.toEntity(goldHarry);
 
@@ -88,46 +72,51 @@ public class StoreService {
     }
 
     //상점 수정
-    @Transactional
-    public void updateStore(String storeId, StoreDto.StoreUpdate updateInfo) {
+    public void updateStore(String token, String storeId, StoreDto.StoreUpdate updateInfo) {
         Store store = storeRepository.findById(Long.valueOf(storeId))
-                .orElseThrow(() -> new NotFoundException("존재하지 않은 상점 입니다."));
+                .orElseThrow(() -> new NotFoundException(NOT_FOUND_STORE));
 
-        StoreDto.StoreInfo storeInfo = updateInfo.getStoreInfo();
+        if (authorityUserRoleUtil.storeVerification(token, store)) {
+            StoreDto.StoreInfo storeInfo = updateInfo.getStoreInfo();
 
-        //상점 이름 검증
-        if (!store.getStoreName().equals(storeInfo.getStoreName()) &&
-                storeRepository.existsByStoreName(storeInfo.getStoreName())) {
-            throw new NotFoundException("이미 존재하는 상점입니다.");
+            //상점 이름 검증
+            if (!store.getStoreName().equals(storeInfo.getStoreName()) &&
+                    storeRepository.existsByStoreName(storeInfo.getStoreName())) {
+                throw new NotFoundException(ALREADY_EXIST_STORE);
+            }
+
+            Address address = addressRepository.findById(Long.valueOf(updateInfo.getAddressId()))
+                    .orElseThrow(() -> new NotFoundException(WRONG_ADDRESS));
+            CommonOption commonOption = commonOptionRepository.findById(Long.valueOf(updateInfo.getCommonOptionId()))
+                    .orElseThrow(() -> new NotFoundException(WRONG_COMMON_OPTION));
+            AdditionalOption additionalOption = additionalOptionRepository.findById(Long.valueOf(updateInfo.getAdditionalOptionId()))
+                    .orElseThrow(() -> new NotFoundException(WRONG_ADDITION_OPTION));
+            GoldHarry goldHarry = goldHarryRepository.findById(Long.valueOf(updateInfo.getGoldHarryId()))
+                    .orElseThrow(() -> new NotFoundException(WRONG_HARRY));
+
+            commonOption.update(updateInfo.getCommonOptionInfo());
+            commonOption.setGoldHarry(goldHarry);
+
+            store.updateStoreInfo(storeInfo);
+            store.setAddress(address);
+            store.setCommonOption(commonOption);
+            store.setAdditionalOption(additionalOption);
+            return;
         }
 
-        Address address = addressRepository.findById(Long.valueOf(updateInfo.getAddressId()))
-                .orElseThrow(() -> new NotFoundException("잘못된 주소 입력입니다."));
-        CommonOption commonOption = commonOptionRepository.findById(Long.valueOf(updateInfo.getCommonOptionId()))
-                .orElseThrow(() -> new NotFoundException("잘못된 기본 옵션 입력입니다."));
-        AdditionalOption additionalOption = additionalOptionRepository.findById(Long.valueOf(updateInfo.getAdditionalOptionId()))
-                .orElseThrow(() -> new NotFoundException("잘못된 추가 옵션 입력입니다."));
-        GoldHarry goldHarry = goldHarryRepository.findById(Long.valueOf(updateInfo.getGoldHarryId()))
-                .orElseThrow(() -> new NotFoundException("잘못된 해리값 입력입니다."));
-
-        commonOption.update(updateInfo.getCommonOptionInfo());
-        commonOption.setGoldHarry(goldHarry);
-
-        store.updateStoreInfo(storeInfo);
-        store.setAddress(address);
-        store.setCommonOption(commonOption);
-        store.setAdditionalOption(additionalOption);
+        throw new NotAuthorityException(NO_ROLE);
     }
 
     //상점 삭제
-    @Transactional
-    public void deleteStore(String storeId) {
+    public void deleteStore(String token, String storeId) {
         Store store = storeRepository.findById(Long.valueOf(storeId))
-                .orElseThrow(() -> new NotFoundException("존재하지 않은 상점 입니다."));
+                .orElseThrow(() -> new NotFoundException(NOT_FOUND_STORE));
 
-        storeRepository.delete(store);
+        if (authorityUserRoleUtil.storeVerification(token, store)) {
+            storeRepository.delete(store);
+            return;
+        }
+        throw new NotAuthorityException(NO_ROLE);
     }
-
-
 
 }
